requirements.txt
requests
beautifulsoup4
google-generativeai
gTTS
python-dotenv
